Add files via upload
One of the main problems with 3d scans are the feathering that occurs within the middle of the scan. Here, we quantify this feathering by taking samples at 20 specific y values in the surface matrix that are in the middle 2/3 of the image. Then with these rows of the surface matrix we measure the proportion of missing values in each of those 20 rows. We then take the median proportion of missing values of those rows, which will give us a good measure to quantify the amount of feathering in a 3d scan.